requirements.txt
livekit-agents>=0.10.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
crewai>=0.11.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
langchain-groq>=0.1.0
groq>=0.5.0
streamlit>=1.28.0
pydantic>=2.0.0
typing-extensions>=4.8.0
openai>=1.0.0
langchain-openai>=0.1.0